Refactor and rename
