+++ apps/discord-bot/src/commands/settings/open-channel-settings-menu.tsx
@@ -1,40 +1,49 @@
+import { ChannelSettingsMenu } from "~discord-bot/components/channel-settings-menu";
 import { ApplyOptions } from "@sapphire/decorators";
 import { Command, container, type ChatInputCommand } from "@sapphire/framework";
 import { callAPI, makeEphemeralErrorHandler } from "~discord-bot/utils/trpc";
-import { SlashCommandBuilder, type ChatInputCommandInteraction } from "discord.js";
+import {
+  SlashCommandBuilder,
+  PermissionsBitField,
+  type ChatInputCommandInteraction,
+} from "discord.js";
 import React from "react";
 import { ephemeralReply } from "~discord-bot/utils/utils";
-import { getDefaultUserServerSettingsWithFlags } from "@answeroverflow/db";
+import { ChannelWithFlags, getDefaultChannelWithFlags } from "@answeroverflow/db";
 import { TRPCError } from "@trpc/server";
 import { createMemberCtx } from "~discord-bot/utils/context";
-
+import { toAOChannel } from "~discord-bot/utils/conversions";
 import { guildTextChannelOnlyInteraction } from "~discord-bot/utils/conditions";
-import { ManageAccountMenu } from "~discord-bot/components/manage-account-menu";
 
 @ApplyOptions<Command.Options>({
-  name: "manage-account",
-  description: "Manage how Answer Overflow interacts with your account",
+  name: "channel-settings",
+  description: "Configure channel settings",
   runIn: ["GUILD_ANY"],
+  requiredUserPermissions: ["ManageGuild"],
 })
-export class OpenManageAccountMenuCommand extends Command {
+export class ChannelSettingsCommand extends Command {
   public override registerApplicationCommands(registry: ChatInputCommand.Registry) {
     registry.registerChatInputCommand(
       new SlashCommandBuilder()
         .setName(this.name)
         .setDescription(this.description)
         .setDMPermission(false)
+        .setDefaultMemberPermissions(PermissionsBitField.resolve("ManageGuild")),
+      {
+        idHints: ["1048055954618454026"],
+      }
     );
   }
 
   public override async chatInputRun(interaction: ChatInputCommandInteraction) {
-    await guildTextChannelOnlyInteraction(interaction, async ({ guild, member }) => {
+    await guildTextChannelOnlyInteraction(interaction, async ({ channel, member }) => {
+      const parentId = channel.isThread() ? channel.parentId : channel.id;
+      if (!parentId) return; // TODO: Send message to user
+
       await callAPI({
         async ApiCall(router) {
           try {
-            return await router.userServerSettings.byId({
-              userId: member.id,
-              serverId: guild.id,
-            });
+            return await router.channels.byId(parentId);
           } catch (error) {
             if (error instanceof TRPCError && error.code == "NOT_FOUND") {
               return null;
@@ -45,13 +54,12 @@ export class OpenManageAccountMenuCommand extends Command {
         },
         Ok(result) {
           if (!result) {
-            result = getDefaultUserServerSettingsWithFlags({
-              userId: interaction.user.id,
-              serverId: guild.id,
-            });
+            result = getDefaultChannelWithFlags(toAOChannel(channel));
           }
           // TODO: Maybe assert that it matches that spec instead of casting
-          const menu = <ManageAccountMenu initalSettings={result} />;
+          const menu = (
+            <ChannelSettingsMenu channel={channel} settings={result as ChannelWithFlags} />
+          );
           ephemeralReply(container.reacord, menu, interaction);
         },
         ...makeEphemeralErrorHandler(interaction),
